# merged-line-openapi.yaml
openapi: 3.0.0
info:
  title: Merged API
  version: 1.0.0
  description: Merged OpenAPI specification
paths:
...
  /v2/bot/message/{messageId}/content:
    get:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#get-content
      servers:
      - url: https://api-data.line.me
      tags:
      - messaging-api-blob
      operationId: getMessageContent
      description: Download image, video, and audio data sent from users.
      parameters:
      - name: messageId
        in: path
        required: true
        description: Message ID of video or audio
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                description: blob response
                type: string
                format: binary
  /v2/bot/message/{messageId}/content/preview:
    get:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#get-image-or-video-preview
      servers:
      - url: https://api-data.line.me
      tags:
      - messaging-api-blob
      operationId: getMessageContentPreview
      description: Get a preview image of the image or video
      parameters:
      - name: messageId
        in: path
        required: true
        description: Message ID of image or video
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: binary
                description: blob response
  /v2/bot/message/{messageId}/content/transcoding:
    get:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#verify-video-or-audio-preparation-status
      servers:
      - url: https://api-data.line.me
      tags:
      - messaging-api-blob
      operationId: getMessageContentTranscodingByMessageId
      description: Verify the preparation status of a video or audio for getting
      parameters:
      - name: messageId
        in: path
        required: true
        description: Message ID of video or audio
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMessageContentTranscodingResponse'
              example:
                status: processing
...
  /v2/bot/richmenu/{richMenuId}/content:
    get:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#download-rich-menu-image
      servers:
      - url: https://api-data.line.me
      tags:
      - messaging-api-blob
      operationId: getRichMenuImage
      description: Download rich menu image.
      parameters:
      - name: richMenuId
        in: path
        required: true
        description: ID of the rich menu with the image to be downloaded
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: binary
                description: blob response
    post:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#upload-rich-menu-image
      servers:
      - url: https://api-data.line.me
      tags:
      - messaging-api-blob
      operationId: setRichMenuImage
      description: Upload rich menu image
      parameters:
      - name: richMenuId
        in: path
        required: true
        description: The ID of the rich menu to attach the image to
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              description: blob request
              type: string
              format: binary
      responses:
        '200':
          description: OK
...
  /v2/bot/audienceGroup/upload/byFile:
    post:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#create-upload-audience-group-by-file
      servers:
      - url: https://api-data.line.me
      tags:
      - manage-audience-blob
      operationId: createAudienceForUploadingUserIds
      description: Create audience for uploading user IDs (by file).
      requestBody:
        content:
          multipart/form-data:
            schema:
              description: Request for createAudienceForUploadingUserIds
              type: object
              required:
              - file
              properties:
                description:
                  type: string
                  maxLength: 120
                  description: 'The audience''s name. This is case-insensitive, meaning
                    AUDIENCE and audience are considered identical.

                    Max character limit: 120

                    '
                isIfaAudience:
                  type: boolean
                  description: 'To specify recipients by IFAs: set `true`.

                    To specify recipients by user IDs: set `false` or omit isIfaAudience
                    property.

                    '
                uploadDescription:
                  type: string
                  description: 'The description to register for the job (in `jobs[].description`).

                    '
                file:
                  type: string
                  format: binary
                  description: 'A text file with one user ID or IFA entered per line.
                    Specify text/plain as Content-Type. Max file number: 1

                    Max number: 1,500,000

                    '
            encoding:
              file:
                contentType: text/plain
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAudienceGroupResponse'
              example:
                audienceGroupId: 1234567890123
                createRoute: MESSAGING_API
                type: UPLOAD
                description: audienceGroupName_01
                created: 1613700237
                permission: READ_WRITE
                expireTimestamp: 1629252237
                isIfaAudience: false
    put:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#update-upload-audience-group-by-file
      servers:
      - url: https://api-data.line.me
      tags:
      - manage-audience-blob
      operationId: addUserIdsToAudience
      description: Add user IDs or Identifiers for Advertisers (IFAs) to an audience
        for uploading user IDs (by file).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              description: Request of the addUserIdsToAudience
              type: object
              required:
              - file
              properties:
                audienceGroupId:
                  type: integer
                  format: int64
                  description: The audience ID.
                uploadDescription:
                  type: string
                  description: The description to register with the job
                file:
                  type: string
                  format: binary
                  description: 'A text file with one user ID or IFA entered per line.
                    Specify text/plain as Content-Type. Max file number: 1

                    Max number: 1,500,000

                    '
            encoding:
              file:
                contentType: text/plain
      responses:
        '202':
          description: OK
...
  /shop/v3/mission:
    post:
      externalDocs:
        url: https://developers.line.biz/en/reference/partner-docs/#send-mission-stickers-v3
      tags:
      - shop
      operationId: missionStickerV3
      description: Sends a mission sticker.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionStickerRequest'
        required: true
      responses:
        '200':
          description: Returns status code 200 and an empty response body.
...
components:
  schemas:
...
    GetMessageContentTranscodingResponse:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#verify-video-or-audio-preparation-status
      description: Transcoding response
      required:
      - status
      type: object
      properties:
        status:
          type: string
          description: 'The preparation status. One of:


            `processing`: Preparing to get content.

            `succeeded`: Ready to get the content. You can get the content sent by
            users.

            `failed`: Failed to prepare to get the content.

            '
          enum:
          - processing
          - succeeded
          - failed
...
    CreateAudienceGroupResponse:
      externalDocs:
        url: https://developers.line.biz/en/reference/messaging-api/#create-upload-audience-group
      type: object
      description: Create audience for uploading user IDs (by JSON)
      properties:
        audienceGroupId:
          type: integer
          format: int64
          description: The audience ID.
        createRoute:
          type: string
          enum:
          - MESSAGING_API
          description: 'How the audience was created.


            `MESSAGING_API`: An audience created with Messaging API.

            '
        type:
          $ref: '#/components/schemas/AudienceGroupType'
        description:
          type: string
          description: The audience's name.
        created:
          type: integer
          format: int64
          description: When the audience was created (in UNIX time).
        permission:
          type: string
          enum:
          - READ
          - READ_WRITE
          description: 'Audience''s update permission. Audiences linked to the same
            channel will be READ_WRITE.


            `READ`: Can use only.

            `READ_WRITE`: Can use and update.

            '
        expireTimestamp:
          type: number
          format: int64
          description: 'Time of audience expiration. Only returned for specific audiences.

            '
        isIfaAudience:
          type: boolean
          description: 'The value indicating the type of account to be sent, as specified
            when creating the audience for uploading user IDs. One of:


            `true`: Accounts are specified with IFAs.

            `false` (default): Accounts are specified with user IDs.

            '
    AudienceGroupType:
      description: Audience group type
      type: string
      enum:
      - UPLOAD
      - CLICK
      - IMP
      - CHAT_TAG
      - FRIEND_PATH
      - RESERVATION
      - APP_EVENT
      - VIDEO_VIEW
      - WEBTRAFFIC
      - IMAGE_CLICK
      - RICHMENU_IMP
      - RICHMENU_CLICK
...
    MissionStickerRequest:
      externalDocs:
        url: https://developers.line.biz/en/reference/partner-docs/#send-mission-stickers-v3
      description: Send mission stickers (v3)
      required:
      - to
      - productId
      - productType
      - sendPresentMessage
      type: object
      properties:
        to:
          type: string
          description: Destination user ID
        productId:
          type: string
          description: Package ID for a set of stickers
        productType:
          type: string
          description: '`STICKER`'
          example: STICKER
        sendPresentMessage:
          type: boolean
          description: '`false`'
          example: false
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: Channel access token
    basicAuth:
      type: http
      scheme: basic