# src/Liny.Core/Controllers/ShopApi.cs
/*
 * Merged API
 *
 * Merged OpenAPI specification
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using System.Text.Json;
using Liny.Core.Attributes;
using Liny.Core.Models;

namespace Liny.Core.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ShopApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Sends a mission sticker.</remarks>
        /// <param name="missionStickerRequest"></param>
        /// <response code="200">Returns status code 200 and an empty response body.</response>
        [HttpPost]
        [Route("/shop/v3/mission")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("MissionStickerV3")]
        public virtual IActionResult MissionStickerV3([FromBody]MissionStickerRequest missionStickerRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }
    }
}